Fix genesis user creation in dev
The create react app was not correctly proxying request to the `/setup` endpoint which the
frontend uses to decide to show the "genesis user" flow on first startup.

It added a trailing slash to the proxied request and Play was fussy and complained that `/setup/`
didn't exist (because it's only `/setup` in the routes).

I think this might have broken via a transitive upgrade to http-proxy, used by http-proxy-middleware.
At least https://github.com/chimurai/http-proxy-middleware/issues/1016 suggests this is this case
and http-proxy is not maintained.