Workspace name column ++60px
Content in this column is often truncated, especially now that we have summaries about number of folders and files within a folder.  The column is so much more important than any other column that I don't mind that this will push "Processing stage" out of view if people use a small browser window.