Add "Copy link to folder" to contextual menu
We're updating the URL to reflect a particular position in a workspace. But this still requires the user to go to the location bar, copy that text (correctly), and then message it to a colleague. To facilitate that this change adds an option to the contextual menu to get the URL of that location.